% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA}
\alias{ASCA}
\alias{print.asca}
\alias{summary.asca}
\title{ASCA}
\usage{
ASCA(d, x, f, comb = NULL, glm_par = vector(mode = "list", length = 0))

\method{print}{asca}(x)

\method{summary}{asca}(x)
}
\arguments{
\item{d}{a data frame with the experimental design, which associate each sample to the design factors}

\item{x}{a matrix with the data measured on the different samples}

\item{f}{a string with the formula which will be used for the \code{glm} decomposition}

\item{comb}{a character vector specifying which terms of the decomposition should be combined}

\item{glm_par}{a list with the parameters to be passed to the \code{glm} call}
}
\value{
An object of class \code{asca}. Actually a list holding the results:
\itemize{
\item of the \code{glm} decomposition
\item of the \code{svd} of the factor matrices
\item of the validation of the object (if present)
}
}
\description{
This function performs the full ASCA decomposition and acts as a constructor of the ASCA S3 object
}
\details{
The ASCA decomposition of a data matrix is performed by using Generalized Linear Models to perform
the estimation of the expected values. the use of GLM's allows the extension of the method to non normal data
and unbalanced designs
}
\note{
There are specific methods to handle and extract information from the \code{asca} object
\itemize{
\item \code{print}
\item \code{summary}
\item \link{scores}
\item \link{loadings}
\item \link{varimp}
\item \link{plot.asca}
\item \link{validate}
}
}
\examples{
## load the data
data("synth_count_data")

## perform the ASCA decomposition
ASCA_test <- ASCA(
d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
glm_par = list(family = poisson())
)

## decomposition with a combination of treatment and interaction
ASCA_test_comb <- ASCA(
d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
comb = c("treatment","time:treatment"),
glm_par = list(family = poisson())
)

## Print a summary of the decomposition
print(ASCA_test)
summary(ASCA_test)

}
