% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA_decompose.R
\name{ASCA_decompose}
\alias{ASCA_decompose}
\title{ASCA decomposition}
\usage{
ASCA_decompose(d, x, f)
}
\arguments{
\item{d}{a data.frame/matrix of factors specifying the design}

\item{x}{a data.fram/matrix with the response (n samples x m variables)}

\item{f}{the formula representing the decomposition}
}
\value{
A list with the following elements:
\itemize{
 \item{decomposition: }{a 3d decomposition array}
 \item{svds: }{the SVD of the ASCA terms}
 \item{exp_variation: }{the initial data frame }
 \item{d: }{the data.frame/matrix of factors specifying the design}
 \item{x: }{the data.fram/matrix with the response (n samples x m variables)}
 \item{f: }{the formula representing the decomposition}
}
}
\description{
Perform the ASCA decomposition of a data matrix according to a given experimental design
}
\details{
The decompusition is calculated by recursively applying lm/glm with the correct cotrasts (sum)
}
