% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_do_bootstrap}
\alias{ASCA_do_bootstrap}
\title{ASCA_do_bootstrap}
\usage{
ASCA_do_bootstrap(d, x)
}
\arguments{
\item{d}{a dataframe with the design}

\item{x}{the data matrix}
}
\value{
a list containing the initial design and the bootstrap data matrix
}
\description{
The function performs a "design" aware bootstrap of a data matrix.
}
\examples{

}
