% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_do_jacknife}
\alias{ASCA_do_jacknife}
\title{ASCA_do_jacknife}
\usage{
ASCA_do_jacknife(d, x, nout = 1)
}
\arguments{
\item{d}{a dataframe with the design}

\item{x}{the data matrix}

\item{nout}{the number of samples for left out for each design level}
}
\value{
a list containing the initial design and the reduced data matrix
}
\description{
The function performs a "design" aware jacknifing of a data matrix.
}
\examples{

}
