% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_fix_sign}
\alias{ASCA_fix_sign}
\title{ASCA_fix_sign}
\usage{
ASCA_fix_sign(A, B)
}
\arguments{
\item{A}{the reference vector}

\item{B}{the second vector}
}
\value{

}
\description{
The function flips the sign of B if the scalar product with A is bigger than 90 degrees
}
\examples{

}
