% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_get_varimp}
\alias{ASCA_get_varimp}
\title{ASCA_get_varimp}
\usage{
ASCA_get_varimp(termlist, ftrans = identity)
}
\arguments{
\item{termlist}{The list of matrices holding the results of an ASCA decomposition}

\item{ftrans}{The function which will be applied to the input before calculating the norm}
}
\value{
a matrix with the variable importance for each term (see \link[gASCA]{ASCA_decompose})
}
\description{
The function calculates the variable importance from the decomposition array. 
See \link[gASCA]{ASCA_decompose} for additional details on how the variable importance is calculated
}
