% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA_decompose.R
\name{ASCA_mnorm}
\alias{ASCA_mnorm}
\title{Matrix norm for ASCA}
\usage{
ASCA_mnorm(m)
}
\arguments{
\item{m}{a matrix of numbers}
}
\value{
the squared Frobenius norm
}
\description{
Calculate the squared norm of a matrix as the square of the Frobenius norm
}
\details{
The squared matrix norm is used to quantify the percentage of variation explained by each ASCA component
}
