% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_norm}
\alias{ASCA_norm}
\title{ASCA_norm}
\usage{
ASCA_norm(m, ftrans = identity)
}
\arguments{
\item{m}{the matrix/vector}

\item{ftrans}{the function which will be applied to the input before calculating the norm}
}
\value{
the norm of the matrix/vector
}
\description{
The function calculates the norm of a vector (standatd norm) or of a matrix (L2 norm) 
after the transformation by f_trans
}
