% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_permutation}
\alias{ASCA_permutation}
\title{ASCA_permutation}
\usage{
ASCA_permutation(asca, turns = 100, qt = 0.95)
}
\arguments{
\item{asca}{an object of class \code{ASCA_decomposition}}

\item{turns}{the number of permutations}

\item{qt}{the empirical quantile}
}
\value{
A list with the following elements
\itemize{
\item \code{R2_qt}: a matrix with the quantiles for the pseusoR2 of the individual variables
\item \code{L2_qt}: a matrix with the quantiles for the L2 norm of the individual terms
\item \code{varimp_qt}: a list holding the matrices of quantiles for each variable in each term
}
}
\description{
the function implements a design permutation strategy to validate the results of an ASCA decomposition
}
\examples{

## load the data
data("synth_count_data")

## perform the ASCA decomposition
dec_test <- ASCA_decompose( d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
glm_par = list(family = poisson())
)

## validate the outcomes

val_out <- ASCA_permutation(dec_test, 10)


}
