% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_permute_design}
\alias{ASCA_permute_design}
\title{ASCA_permute_design}
\usage{
ASCA_permute_design(d)
}
\arguments{
\item{d}{a design data.frame}
}
\value{
a data.frame
}
\description{
The function performs a random permutation of a design data frame.
All columns are permuted independently
}
