% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA_decompose.R
\name{ASCA_svd}
\alias{ASCA_svd}
\title{Decompose the terms ASCA}
\usage{
ASCA_svd(arr, eigthresh = 0.001, returnloadings = TRUE)
}
\arguments{
\item{arr}{a 3d array containing the terms of the decomposition. See details.}

\item{eigthresh}{the threshold (relative importance) used to discard low importance components. The default value is 0.1 "\%" of the sum off the eigenvalues}

\item{returnloadings}{should the loadings of svd be returned}
}
\value{
A list with the following elements:
\itemize{
 \item{loadings: }{the loadings of the initial variables in each term}
 \item{varimp: }{the importance of each initial variable for the decomposition of each term}
 \item{variances: }{the variance cexplained by each decomposition term, calculated from the SVD eigenvalues}
}
}
\description{
The function performs the svd decomposition of the different ASCA terms
}
\details{
The input array is 3D. The first dimensions is the row of the initial matrix (samples).
The second dimension encodes the ASCA decomposition terms. The third one represents the columns of the intial
data matrix, i.e. the original variables.

The explained variance is calculated as the squared sum of the "relevant" terms of the d vector containing the singular values obtained from svd
}
\examples{

}
