% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_svd_response}
\alias{ASCA_svd_response}
\title{ASCA_svd_response}
\usage{
ASCA_svd_response(arr, invlink = identity)
}
\arguments{
\item{arr}{the multidimensional (2D or 3D) array. In 2D case rows are samples and columns variables. 
In the 3D case (the output of `ASCA_decompose`) organized as samples x terms x variables 
the SVD is performed along the 2nd dimension, aka one SVD per term.
If the matrix has rank 1 the decompoition is set to NULL}

\item{invlink}{the function to be applied to the scores of the terms that are not NULL or error}
}
\value{
a list with the results of the SVD decomposition. 
 * scores: a list with the scores for each term, in the response
 * loadings: a list with the scores for each term
 * varimp: a list with the importance of each variable for each term
 * variances: a list with the variance of each eigenvector for each term
}
\description{
ASCA_svd_response
}
