% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA_decompose.R
\name{ASCA_validate}
\alias{ASCA_validate}
\title{ASCA_validate}
\usage{
ASCA_validate(asca, npermutations = 500)
}
\arguments{
\item{asca}{the output of the ASCA_decompose}

\item{npermutations}{the number of permutations used to construct the null distribution}
}
\value{

}
\description{
apply a permutation approach to validate the significance of the different terms of the ASCA decomposition
}
\details{
The significance of each term of the ASCA decomposition
is evaluated by calculating the empirical distribution of the explained variance by recursive permutation of the
design factors. Each decomposition term is considered significance if it accounts for a fraction of variance
larger than the one that would be obtained by chance by a random assignment of the samples to the design factors
}
\examples{

}
