% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.varimp.R
\name{ASCA_validate.varimp}
\alias{ASCA_validate.varimp}
\title{Title}
\usage{
ASCA_validate.varimp(asca, npermutations = 100, quantile = 0.95)
}
\arguments{
\item{asca}{the output of the ASCA_decompose}

\item{npermutations}{the number of permutations used to construct the null distribution}

\item{quantile}{the upper quantile of the random variable importance table}
}
\value{
a list with the following item
\itemize{
 \item{random.varimp.array: }{an array (samples x factors x npermutations) with the random variables importance generated by the ASCA decomposition of permuted data }
 \item{upperquantile: }{the upper quantile of the random variable importance}
 \item{important.var: }{comparison of the observed variable importance with the upper quantile from the randomised variable importance }
}
}
\description{
Title
}
\details{
Entries in the important.var item are either "NA" for variables  whose importance is below the upper quantile of randomised data
or the actual variable importance values for those > than the randomised variable importance value
}
\examples{



}
