% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA_validate2.R
\name{ASCA_validate2}
\alias{ASCA_validate2}
\title{ASCA validation2}
\usage{
ASCA_validate2(asca, npermutations = 100, upquantile = 0.95)
}
\arguments{
\item{asca}{The object retured by the ASCA_decompose}

\item{npermutations}{Number of permutations to be used to construct the null distribution of ASCA decomposition terms}

\item{upquantile}{The upper quantile limits of the null distribution of ASCA terms, used for testing significance}
}
\value{
Returns a list with the following elements:
\itemize{
 \item{random_variances: }{a matrix with null values of explained variance by each factor}
 \item{upper.quantile.variances: }{the upper quantile of the null distribution of explained variance}
 \item{random_varimp.array: }{an array of null values of variable importance; the last dimension of the array equals npermutation}
 \item{upper.quantile.varimp: }{the observed variable importance, reporting only variable-factor cases where values > null upperquantile}
 }
}
\description{
Function to validate the ASCA decomposition; apply a permutation approach to validate the significance of the different terms of the ASCA decomposition
}
\details{
The significance of each term of the ASCA decomposition
is evaluated by calculating the empirical distribution of the explained variance and variable importance 
by recursive permutation of the design factors. Each decomposition term is considered significance if it accounts for a fraction of variance
larger than the one that would be obtained by chance by a random assignment of the samples to the design factors
}
\examples{


}
