% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{get_varimp}
\alias{get_varimp}
\title{get_varimp}
\usage{
get_varimp(termarray)
}
\arguments{
\item{termarray}{}
}
\value{

}
\description{
The function calculates the variable importance from the decomposition array.
}
\examples{

}
