% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{getValidation}
\alias{getValidation}
\title{getValidation.asca}
\usage{
getValidation(x)
}
\arguments{
\item{x}{the asca object}
}
\value{
A list with the resampling quantiles for the loadings and the variable importance
}
\description{
Extracts the results of the validation from the `asca` object
}
