% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{loadings}
\alias{loadings}
\title{loadings}
\usage{
loadings(x)
}
\arguments{
\item{x}{an object of class asca}
}
\value{
A list of matrices with the loadings of the decomposition. The number of eigenvectors 
is determined by the number of levels for each factor
}
\description{
Extracts the loadings of the different terms from the `asca` object
}
