% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA_validate2.R
\name{plot.validate.var}
\alias{plot.validate.var}
\title{Plot ASCA validate}
\usage{
\method{plot}{validate.var}(asca, asca_validate, var, fact)
}
\arguments{
\item{asca}{The object retured by the ASCA_decompose}

\item{asca_validate}{The object returned by the ASCA_validate2}

\item{var}{The variable that needs to plot (called by number)}

\item{fact}{The factor for which the variable importance is calculated (called by number)}
}
\value{

}
\description{
Plot ASCA validate
}
\details{
The plot is used to visually explore the deviation of the observed variable importance (for a given factor) relative to its null distribution.
Each variable and factor should be called in the function
}
\examples{



}
