% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.varimp.R
\name{plot.validate.varimp}
\alias{plot.validate.varimp}
\title{Title}
\usage{
\method{plot}{validate.varimp}(asca, validate.varimp, var, fact)
}
\arguments{
\item{asca}{the output of ASCA_decompose}

\item{validate.varimp}{the output of ASCA_validate.varimp}

\item{var}{the variable [number] to plot}

\item{fact}{the factor [number] to plot}
}
\value{
an histogram of the random variable importance values and the location of the observed value
}
\description{
Title
}
\examples{

}
