% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{scores}
\alias{scores}
\title{scores}
\usage{
scores(x)
}
\arguments{
\item{x}{an object of class asca}
}
\value{
A list of matrices with the scores of the decomposition. The number of eigenvectors 
is determined by the number of levels for each factor
}
\description{
Extract the scores from an `asca` object
}
