% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_count_data}
\alias{synth_count_data}
\title{Synthetic count data}
\format{
a list with two components:
\describe{
\item{counts}{a matrix with the counts (45x10)}
\item{design}{a data.frame with the designh factors}
...
}
}
\usage{
synth_count_data
}
\description{
A dataset mimicking the results of a counting experiments of 10 variables over 45 samples.

Design:
\itemize{
\item 5 timepoints;
\item 3 treatments ;
\item 10 replicates
}

The variables were created from a two latent variable model which were "projected" on the the final variables. In details
\itemize{
\item s1,s2,s3,s4 respond to the "time" factor. s1 and s2 contribute to the first latent factor, s3 and s4 to the second
\item s5,s6  respond to the "treatment" factor
\item s1 and s2 respond also to the interaction term "time:treatment". In other words, for them the effect of treatment depends on time.
}
}
\keyword{datasets}
