% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca_validate.R
\name{validate}
\alias{validate}
\title{validate}
\usage{
validate(x, method, turns, nout, qt)
}
\arguments{
\item{x}{An asca object}

\item{method}{A character vector which specify the type of validation strategy. Either "jacknifing" (the default) or
"bootstrap".}

\item{turns}{How many independent resampling should be performed}

\item{nout}{How many samples will be "left" out in the case of jacknifing}

\item{qt}{A vector with the quantiles which will be calculated on the resamplings}
}
\value{
an object of class \code{asca} containing also the results of the validation.
}
\description{
This function implements a resampling strategy (based on bootstrap or jacknifing) to validate the results
of an ASCA decomposition in terms of loadings and variable importance
}
\details{
Validation is performed by evaluating the variability of the estimates of the variable importance and
of the loadings upon a bootstrapping/jacknifing which preserve the experimental design. The results are
presented in form of the empirical quantiles of the validated parameters.
}
\examples{

## load the data
data("synth_count_data")

## perform the ASCA decomposition
ASCA_test <- ASCA(
d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
glm_par = list(family = poisson()))

## perform validation
ASCA_test_v <- validate(ASCA_test,
turns = 100,
)

## the results are included in the asca object inside the validation element

}
