% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{varimp}
\alias{varimp}
\title{varimp}
\usage{
varimp(x)
}
\arguments{
\item{x}{the asca object}
}
\value{
A matrix with the variable importance for each factor of the decomposition
}
\description{
Extracts the loadings of the different terms from the `asca` object
}
\details{
The variable importance is calculated by weighting the absolute value of the loadings of the 
singular values by the relative importance of the eigenvalues
}
