% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_combine_terms}
\alias{ASCA_combine_terms}
\title{ASCA_combine_terms}
\usage{
ASCA_combine_terms(arr3d, comb)
}
\arguments{
\item{arr3d}{}

\item{comb}{}
}
\value{

}
\description{
ASCA_combine_terms
}
\examples{

}
