% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_decompose}
\alias{ASCA_decompose}
\title{ASCA_decompose}
\usage{
ASCA_decompose(
  d,
  x,
  f,
  glm_par = vector(mode = "list", length = 0),
  res_type = "response"
)
}
\arguments{
\item{d}{a data.frame/matrix with the design}

\item{x}{a data.frame/matrix of numeric values to be decomposed}

\item{f}{a string holding the formula of the decomposition}

\item{glm_par}{a list with the parameters to be passed to the \code{glm} call}

\item{res_type}{the types of GLM residuals}
}
\value{
a list with the full outcomes of the decomposition with the following elements
\itemize{
\item decomposition: a list holding the results of the decomposition
\item mu: a vector with the constant terms of the univariate models
\item residuals: a matrix holding the model residuals. Their type is stored in the \code{res_type} element.
\item prediction:  the matrix with the predicted values in the linear predictor space
\item pseudoR2: a parameter to assess the goodness of fit for the model on each variable.
\item glm_par: a list with the parameters used for modelling
\item res_type: the type of residuals
\item varimp: the importance of the individual variables in the decomposition terms
\item terms_L2: the L2 norm of the individual terms back transformed in the response space
\item d: a data.frame with the design
\item x: a data.frame with the initial data
\item f: the string defining the decomposition
\item combined: a vector holding the combined terms ()
\item invlink: the inverse of the link function used in the glm fitting
}
}
\description{
This function performs the full ASCA decomposition
}
\details{
The ASCA decomposition of a data matrix is performed by using Generalized Linear Models
to estimate univariate expected values. The use of GLM's allows the extension of the method to non normal data
and unbalanced designs.
This function performs only the decomposition without the SVD, which have to be performed by \code{ASCA_svd}.
The level of fit for each variable is assessed calculating the \code{pseudoR2}, which is defined as:

\deqn{1-residual_deviance/null_deviance}

The variable importance element stores a measure of the importance of each variable \eqn{c} for each term calculated
as the norm of each column. L2 norm is also calculated for each decomposition term.

It is important to highlight that in the case of count data with large fraction of zeroes the variable importance and the
term L2 norms cannot be considered as reliable measure of importance because in presence of log links very low expected values are associated to
very large negative values in the linear predictor space.
}
\examples{
## load the data
data("synth_count_data")

## perform the ASCA decomposition
dec_test <- ASCA_decompose(
d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
glm_par = list(family = poisson())
)


}
