% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_decompose}
\alias{ASCA_decompose}
\title{ASCA_decompose}
\usage{
ASCA_decompose(
  d,
  x,
  f,
  glm_par = vector(mode = "list", length = 0),
  res_type = "pearson"
)
}
\arguments{
\item{d}{a data.frame/matrix with the design}

\item{x}{a data.frame/matrix of numeric values to be decomposed}

\item{f}{a string holding the formula of the decomposition}

\item{glm_par}{a list with the parameters to be passed to the \code{glm} call}

\item{res_type}{the types of GLM residuals}
}
\value{
a list with the full outcomes of the decomposition with the following elements
\itemize{
\item decomposition: the 3D array holding the results of the decomposition
\item deviance: the array containing the deviance of the models
\item glm_par: a list with the parameters used for modeling
\item res_type: the type of residuals
}
}
\description{
This function performs the full ASCA decomposition
}
\details{
The ASCA decomposition of a data matrix is performed by using Generalized Linear Models to perform
the estimation of the expected values. The use of GLM's allows the extension of the method to non normal data
and unbalanced designs.
This function performs only the decomposition without the SVD, which have to be performed by \code{ASCA_svd}
}
\examples{
## load the data
data("synth_count_data")

## perform the ASCA decomposition
ASCA_test <- ASCA_decompose(
d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
glm_par = list(family = poisson())
)


}
