% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_get_varimp}
\alias{ASCA_get_varimp}
\title{ASCA_get_varimp}
\usage{
ASCA_get_varimp(termlist, linkinv)
}
\arguments{
\item{termlist}{The list of matrices holding the results of an ASCA decomposition}

\item{invlink}{The inverse link function to be applied to back transform the vectors}
}
\value{
a matrix with the variable importance for each term (see \link[ASCA]{ASCA_decompose})
}
\description{
The function calculates the variable importance from the decomposition array. 
See \link[ASCA]{ASCA_decompose} for additional details on how the variable importance is calculated
}
