% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_svd}
\alias{ASCA_svd}
\title{ASCA_svd}
\usage{
ASCA_svd(decomposition, ...)
}
\arguments{
\item{decomposition}{the decomposition list produced by  \code{ASCA_decompose}.}

\item{...}{additional parameters to be passed to \code{prcomp}}
}
\value{
a list with the results of the SVD decomposition for each term. Each element is the output of a
\code{prcomp}
}
\description{
ASCA_svd
}
\examples{
## load the data
data("synth_count_data")

## perform the ASCA decomposition
dec_test <- ASCA_decompose(
d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
glm_par = list(family = poisson())
)

## Run the SVD of the terms
svd_test <- ASCA_svd(dec_test$decomposition)
 
}
