% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCA.R
\name{ASCA_trim_vars}
\alias{ASCA_trim_vars}
\title{ASCA_trim_vars}
\usage{
ASCA_trim_vars(decomposition, keep = NULL)
}
\arguments{
\item{decomposition}{a list of decomposition matrices produced by `ASCA_decompose` or `ASCA_combine_terms`}

\item{keep}{a vector of numbers or names indicating the variables which should be kept}
}
\value{
a list of reduced matrices
}
\description{
ASCA_trim_vars
}
