% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_count_data}
\alias{synth_count_data}
\title{Synthetic count data}
\format{
a list with two components:
\describe{
  \item{counts}{a matrix with the counts (45x10)}
  \item{design}{a data.frame with the designh factors}
  ...
}
}
\usage{
synth_count_data
}
\description{
A dataset mimicking the results of a counting experiments of 10 variables over 45 samples.
}
\details{
Design: 

* 5 timepoints; 
* 3 treatments ;
* 3 replicates

The variables were created from a two latent variable model which were "projected" on the the final variables
}
\keyword{datasets}
